feat(network): add dto, entity

+++ WhenComing/Data/Network/APIEndpoint.swift
@@ -11,7 +11,7 @@ enum APIEndpoint {
     // 버스정류소정보 busArrivalInfo
     case getCityCodeList                    // 서비스 가능 지역 검색
     case getStationList(pageNo: Int, cityCode: String, stationName: String)  // 정류소번호 목록조회: 정류장 이름 검색 결과 가져오기
-    case getAroundStationList(pageNo: Int, lat: Float, lng: Float) // 좌표기반근접정류소 목록조회: 주변 정류장 뭐뭐 있는지
+    case getAroundStationList(pageNo: Int, lat: Double, lng: Double) // 좌표기반근접정류소 목록조회: 주변 정류장 뭐뭐 있는지
     case getSttnThrghRouteList(pageNo: Int, cityCode: String, nodeId: String) // 정류소별경유노선 목록조회: 무슨무슨 버스 오는지
     
     // 버스도착정보 busStationInfo
